scal_local: 3rd part: now the code compiles

+++ model/src/w3triamd.F90
@@ -1584,7 +1584,7 @@ END SUBROUTINE
 ! Parameter list
 
      INTEGER, INTENT(IN)            :: IMOD
-     REAL   , INTENT(IN)            :: XTIN, YTIN
+     DOUBLE PRECISION, INTENT(IN)   :: XTIN, YTIN
      INTEGER, INTENT(OUT)           :: itout
      INTEGER, INTENT(OUT)           :: IS(4), JS(4)
      REAL, INTENT(OUT)              :: RW(4)
@@ -1762,7 +1762,7 @@ END SUBROUTINE
 ! Parameter list
 
      INTEGER, INTENT(IN)            :: IMOD, FORCE
-     REAL   , INTENT(IN)            :: XTIN, YTIN
+     DOUBLE PRECISION, INTENT(IN)            :: XTIN, YTIN
      INTEGER, INTENT(OUT)           :: itout
      INTEGER, INTENT(OUT)           :: IS(4), JS(4)
      REAL, INTENT(OUT)              :: RW(4)
